Using Enumerable.Empty() as a falback value for null-source IEnumerable<T> mappings

+++ AgileMapper.UnitTests/SimpleTypeConversion/WhenConvertingToBytes.cs
@@ -294,7 +294,7 @@
             var source = new[] { "9", "8", "7" };
             var result = Mapper.Map(source).ToANew<IEnumerable<byte>>();
 
-            result.ShouldBe(s => (int)s, 9, 8, 7);
+            result.ShouldBe(s => s, 9, 8, 7);
         }
     }
 }
